# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EstoqueClinic API",
    "version": "1.0.0",
    "description": "API REST para gestão de estoque de clínicas estéticas com multi-tenancy",
    "contact": {
      "name": "EstoqueClinic Support",
      "email": "suporte@estoqueclinic.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento Local"
    },
    {
      "url": "https://api.estoqueclinic.com",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de autenticação. Header: Authorization: Bearer {token}"
      }
    },
    "schemas": {
...
      "AestheticProduct": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string"},
          "brand": {"type": "string"},
          "category": {"type": "string", "enum": ["botox", "filler", "biostimulator", "equipment", "consumable"]},
          "subcategory": {"type": "string"},
          "concentration": {"type": "string"},
          "volume_ml": {"type": "number"},
          "units_per_package": {"type": "integer"},
          "anvisa_registry": {"type": "string"},
          "manufacturer": {"type": "string"},
          "active_principle": {"type": "string"},
          "storage_temp_min": {"type": "number"},
          "storage_temp_max": {"type": "number"},
          "shelf_life_months": {"type": "integer"},
          "description": {"type": "string"},
          "usage_instructions": {"type": "string"},
          "contraindications": {"type": "string"},
          "image_url": {"type": "string"},
          "barcode": {"type": "string"},
          "is_controlled": {"type": "boolean"},
          "requires_prescription": {"type": "boolean"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/aesthetic-products": {
      "get": {
        "tags": ["Aesthetic Products"],
        "summary": "Listar produtos estéticos",
        "description": "Lista todos os produtos do catálogo estético",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 20}},
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "category", "in": "query", "schema": {"type": "string", "enum": ["botox", "filler", "biostimulator", "equipment", "consumable"]}},
          {"name": "brand", "in": "query", "schema": {"type": "string"}},
          {"name": "is_active", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos estéticos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/AestheticProduct"}
                    },
                    "pagination": {"type": "object"}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Aesthetic Products"],
        "summary": "Criar produto estético",
        "description": "Adiciona um novo produto ao catálogo",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "brand", "category"],
                "properties": {
                  "name": {"type": "string"},
                  "brand": {"type": "string"},
                  "category": {"type": "string", "enum": ["botox", "filler", "biostimulator", "equipment", "consumable"]},
                  "subcategory": {"type": "string"},
                  "concentration": {"type": "string"},
                  "volume_ml": {"type": "number"},
                  "units_per_package": {"type": "integer"},
                  "anvisa_registry": {"type": "string"},
                  "manufacturer": {"type": "string"},
                  "active_principle": {"type": "string"},
                  "storage_temp_min": {"type": "number"},
                  "storage_temp_max": {"type": "number"},
                  "shelf_life_months": {"type": "integer"},
                  "description": {"type": "string"},
                  "usage_instructions": {"type": "string"},
                  "contraindications": {"type": "string"},
                  "image_url": {"type": "string"},
                  "barcode": {"type": "string"},
                  "is_controlled": {"type": "boolean"},
                  "requires_prescription": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "product": {"$ref": "#/components/schemas/AestheticProduct"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}